+++ my_game_list/__init__.py
@@ -1,2 +1,2 @@
 """Main __init__, contains the application version number."""
-__version__ = (2, 2, 0)
+__version__ = (3, 0, 0)