+++ my_game_list/games/filters/game.py
@@ -21,6 +21,13 @@ class GameFilterSet(filters.FilterSet):
         to_field_name="name",
         queryset=Platform.objects.all(),
     )
+    ordering = filters.OrderingFilter(
+        fields=(
+            ("created_at", "created_at"),
+            ("rank_position", "rank_position"),
+            ("popularity", "popularity"),
+        ),
+    )
 
     class Meta:
         """Meta class for GameFilterSet."""