+++ my_game_list/my_game_list/forms/__init__.py
@@ -1,6 +1,0 @@
-"""This package contains additional functionality for admin forms."""
-from my_game_list.my_game_list.forms.binary_file_upload_field import BinaryFieldWithUpload
-
-__all__ = [
-    "BinaryFieldWithUpload",
-]