+++ my_game_list/my_game_list/validators/__init__.py
@@ -1,0 +1,6 @@
+"""The validators used across the apps."""
+from my_game_list.my_game_list.validators.file_size_validator import FileSizeValidator
+
+__all__ = [
+    "FileSizeValidator",
+]