+++ my_game_list/users/admin/user.py
@@ -1,10 +1,6 @@
 """This module contains the admin model for the User model."""
-from typing import Any, ClassVar
-
 from django.contrib import admin
-from django.db import models
 
-from my_game_list.my_game_list.forms import BinaryFieldWithUpload
 from my_game_list.users.models import User
 
 
@@ -12,11 +8,6 @@ from my_game_list.users.models import User
 class UserAdmin(admin.ModelAdmin[User]):
     """Admin model for the User model."""
 
-    formfield_overrides: ClassVar[dict[type[models.Field[Any, Any]], dict[str, Any]]] = {  # type: ignore[misc]
-        models.BinaryField: {
-            "form_class": BinaryFieldWithUpload,
-        },
-    }
     readonly_fields = ("id", "avatar_tag")
     search_fields = ("id", "username", "email")
     list_filter = ("is_superuser", "is_staff", "is_active", "date_joined")